requirements.txt
django-environ==0.9.0
psycopg2-binary==2.9.5
django==4.1.5
django-rest==0.8.7
djangorestframework==3.14.0
djoser==2.1.0
django_extensions==3.2.1
